Remove navigation mixin
